Fix parent-link click hitbox and center parent toolbar rows

+++ gedbrowserng-frontend/src/app/modules/person/person-parent.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectorRef, Component, Inject, Input, NgZone, OnChanges, OnInit } from '@angular/core';
 
 import { ApiAttribute } from '../../models';
-import { PersonService } from '../../services';
+import { PersonService, UserService } from '../../services';
 import { HasFamily } from '../../interfaces/has-family';
 import { PersonGetter } from './person-getter';
 import { RefreshPerson } from '../../interfaces';
@@ -12,19 +12,41 @@ import { MatIcon } from '@angular/material/icon';
 
 @Component({
     selector: 'app-person-parent',
-    template: `
-@if (person) {
-  <span><a class="name"
+    template: `@if (person) {
+  <a class="name"
     [routerLink]="['/' + dataset + '/persons', person.string]">
-    {{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]</a>
-  <span class="example-fill-remaining-space"></span>
+    {{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]
+  </a>
+}
+@if (person && hasSignedIn()) {
   <span class="hidden">
     <button mat-icon-button matTooltip="Unlink parent" color="warn" (click)="unlink()">
-      <mat-icon matListIcon>link_off</mat-icon></button>
-  </span>
+      <mat-icon matListIcon>link_off</mat-icon>
+    </button>
   </span>
 }`,
-    styles: [],
+    styles: [`
+:host {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  min-width: 0;
+}
+
+:host .name {
+  display: inline;
+  line-height: 24px;
+  text-decoration: none;
+}
+
+:host:hover .hidden {
+  visibility: visible;
+}
+
+:host:focus-within .hidden {
+  visibility: visible;
+}
+`],
     imports: [RouterLink, MatIconButton, MatTooltip, MatIcon]
 })
 export class PersonParentComponent extends PersonGetter implements OnInit, OnChanges {
@@ -33,6 +55,7 @@ export class PersonParentComponent extends PersonGetter implements OnInit, OnCha
   @Input() attribute: ApiAttribute;
 
   constructor(@Inject(PersonService) personService: PersonService,
+    @Inject(UserService) private readonly userService: UserService,
     @Inject(NgZone) zone: NgZone,
     @Inject(ChangeDetectorRef) cdr: ChangeDetectorRef) {
     super(personService, zone, cdr);
@@ -53,5 +76,9 @@ export class PersonParentComponent extends PersonGetter implements OnInit, OnCha
 
   refreshPerson() {
     this.parent.refreshPerson();
+  }
+
+  hasSignedIn() {
+    return !!this.userService.currentUser;
   }
 }
